Moved Destiny Pool to its own FormApplication.  Added dragging bars to the top and bottom of the display.
