+++ styles/starwarsffg.css
@@ -2615,6 +2615,38 @@ img {
   margin: 5px 2px;
 }
 
+#destiny-tracker {
+  box-shadow: none;
+  background: transparent;
+  position: absolute;
+  right: 305px;
+  bottom: 0;
+}
+
+#destiny-tracker .window-content {
+  background: transparent;
+}
+
+#destiny-tracker header.window-header {
+  display: none;
+}
+
+.swffg-destiny-container {
+  right: 305px;
+  bottom: 0px;
+  position: absolute;
+}
+
+.swffg-destiny-bar {
+  height: 5px;
+  width: 100%;
+  background: url(../../../ui/denim075.png) repeat;
+  box-shadow: 0 0 10px #000;
+  box-sizing: content-box;
+  padding: 0 1px;
+  margin: -3px 0;
+}
+
 .swffg-destiny {
   background: url(../../../ui/denim075.png) repeat;
   border: 1px solid #000;
@@ -2625,14 +2657,12 @@ img {
   font-family: "warnock-pro", "sans-serif";
   font-size: 30px;
   height: auto;
-  right: 305px;
-  bottom: 0px;
   line-height: 1.2;
   margin: 8px 8px 6px 0;
   pointer-events: none;
-  position: fixed;
   text-align: center;
-  width: 84px;
+  min-width: 84px;
+  width: 100%;
   display: flex;
   z-index: 70;
 }
