+++ templates/ffg-destiny-tracker.html
@@ -1,0 +1,8 @@
+<div class="swffg-destiny-container">
+  <div class="swffg-destiny-bar"></div>
+  <div class="swffg-destiny" title="Left click to flip, SHIFT+left click to add, CTRL+left click to remove.">
+    <section id="destinyLight" class="destiny-points" data-value="${destinyPool.light}" data-group="dPoolLight">{{destinyPool.light}}<span>{{localize "SWFFG.Lightside" }}</span></section>
+    <section id="destinyDark" class="destiny-points" data-value="${destinyPool.dark}" data-group="dPoolDark">{{destinyPool.dark}}<span>{{localize "SWFFG.Darkside" }}</span></section>
+  </div>
+  <div class="swffg-destiny-bar"></div>
+</div>
